Setup School Terms
I had two approaches in mind to solve this issue the first involved creating a term/semester plugin which would create a term object and that object be selectable whenever creating a course which would automatically set the course start and end date.

The second approach which I went with was just to add a bunch of custom fields at the end of the moodle course creation page. My rationale was the former would be a major time investment and the latter would be quick and easy if this approach is rejected by the client then we would attempt the firat custom plugin method if not then I just saved a whole lot of dev time.